fix: improve stability and performance
- AudioEncoderWorker: add callbackScope, release(), ShortBuffer optimization
- AudioRingBuffer: fix Long-to-Int overflow in getRecentSamples
- AppDatabase: add MIGRATION_1_2 for baselineRms field
- SleepSessionEntity: add baselineRms field
- SleepRecordingService: Android 13+ RECEIVER_NOT_EXPORTED, state reset, GlobalScope in onDestroy
- HomeViewModel/SessionDetailViewModel: use baselineRms from database

+++ app/src/main/java/com/gx/sleep/audio/AudioRingBuffer.kt
@@ -49,7 +49,7 @@ class AudioRingBuffer(
      */
     @Synchronized
     fun getRecentSamples(maxSamples: Int = totalSamples): ShortArray {
-        val availableSamples = minOf(samplesWritten.toInt(), totalSamples)
+        val availableSamples = minOf(samplesWritten, totalSamples.toLong()).toInt()
         val samplesToReturn = minOf(maxSamples, availableSamples)
 
         if (samplesToReturn == 0) return ShortArray(0)
@@ -84,7 +84,7 @@ class AudioRingBuffer(
      */
     @Synchronized
     fun getAvailableSamples(): Int {
-        return minOf(samplesWritten.toInt(), totalSamples)
+        return minOf(samplesWritten, totalSamples.toLong()).toInt()
     }
 
     /**
